+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 _This is the changelog for the entire project_
 
+## 10/10/2024
+
+- Refactored WeightList and Chart to have loading messages
+
 ## 9/28/2024
 
 - Refactored app to use Loading.Text and Clip Loader